backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
torch==2.1.0
transformers==4.35.0
langchain==0.1.0
langchain-community==0.0.10
llama-cpp-python==0.2.20
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0
python-dotenv==1.0.0
pytz==2023.3
pytest==7.4.3
pytest-asyncio==0.21.1
# Integration dependencies
slack-sdk==3.23.0
microsoft-graph-core==0.2.2
google-auth==2.23.4
